localized defaults interface
